db_updater/requirements.txt
requests==2.23.0
python-dotenv==0.13.0
mysql-connector-python==8.0.20